backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
# Specify compatible versions
motor==3.3.1
pymongo==4.6.0
python-multipart==0.0.6
jinja2==3.1.2
anthropic==0.49.0
python-dotenv==1.0.0 